refactor(admin): extract QualityTab into its own file
